example refactored
