+++ example/lib/components/textfield.dart
@@ -12,8 +12,8 @@ Widget textfield() {
             hintText: "Enter your name",
             controller: TextEditingController(),
             filledColor: Colors.white,
-            suffixIcon: Icons.person,
-            // prefixIcon: Icons.person,
+            suffixIcon: Icons.visibility_off,
+            prefixIcon: Icons.lock,
           ),
         ],
       ),
